Started Android Project
Created the project for our application. I created three of the
activities from Gage's mockup: Summary, Account, and Settings. These are
the only three that can navigate to each other. I added list views to
them as well. We need to throw in some faux data in these views to make
the UI look better for our presentation.